More Streamlining

+++ core_functions/details/server_data_fill.py
@@ -6,17 +6,19 @@ async def server_data_fill(ev):
 
 
 async def server_filler_loop(ev):
-    while True:
-        all_guilds = ev.bot.guilds
-        srv_coll = ev.db[ev.db.db_cfg.database].ServerDetails
-        srv_coll.drop()
-        server_list = []
-        for guild in all_guilds:
-            srv_data = {
-                'Name': guild.name,
-                'ServerID': guild.id,
-                'Icon': guild.icon_url or 'https://i.imgur.com/QnYSlld.png'
-            }
-            server_list.append(srv_data)
-        srv_coll.insert_many(server_list)
-        await asyncio.sleep(3600)
+    if not ev.bot.cooldown.on_cooldown(ev.name, 'server_details'):
+        ev.bot.cooldown.set_cooldown(ev.name, 'server_details', 3600)
+        while True:
+            all_guilds = ev.bot.guilds
+            srv_coll = ev.db[ev.db.db_cfg.database].ServerDetails
+            srv_coll.drop()
+            server_list = []
+            for guild in all_guilds:
+                srv_data = {
+                    'Name': guild.name,
+                    'ServerID': guild.id,
+                    'Icon': guild.icon_url or 'https://i.imgur.com/QnYSlld.png'
+                }
+                server_list.append(srv_data)
+            srv_coll.insert_many(server_list)
+            await asyncio.sleep(300)
